requirements.txt
dotenv
boto3
langchain
pypdf
langchain-community
langchain-pinecone
pinecone-client